requirements.txt
flask
mongoengine
python-telegram-bot
tldextract
psycopg2-binary
python-dotenv